Nanjing problem found

+++ runner.py
@@ -595,7 +595,8 @@ def main():
     if ro_mode == 'file' or ro_mode == '+file':
         ro_file.close()
 
-    os.remove(node_file)
+    if(node_file=="auto")
+        os.remove(node_file)
 
 
 if __name__ == '__main__':
